Implement hero facing direction using FRP sorta
This is a bit buggy: it folds keyboard input into a 4-directional
impulse structure and then maps that to facing direction. But lifting
the last key makes impulse 0, and the default direction is down, so the
character resets direction annoyingly whenever all keys are released.
Need to figure out a better way; either impulse can remain if the last
key is released, or direction needs to be a fold instead of a map...